blinky mais ou menos perseguindo o player

+++ pacman-sin141/clyde.h
@@ -8,7 +8,7 @@ public:
 	Clyde(int x, int y);
 	~Clyde();
 	virtual void renderGhost(ALLEGRO_BITMAP* img, int sprite);
-	virtual void chasePacman(std::vector<std::vector<char>> mapa, float pacmanX, float pacmanY);
+	virtual void chasePacman(std::vector<std::vector<char>> mapa, int pacmanX, int pacmanY);
 };
 
 #endif // !CLYDE_H
