+++ pacman-sin141/inky.h
@@ -7,7 +7,7 @@ public:
 	Inky(int x, int y);
 	~Inky();
 	virtual void renderGhost(ALLEGRO_BITMAP* img, int sprite);
-	virtual void chasePacman(std::vector<std::vector<char>> mapa, float pacmanX, float pacmanY);
+	virtual void chasePacman(std::vector<std::vector<char>> mapa, int pacmanX, int pacmanY);
 };
 
 #endif // !INKY_H
